aisb-sg/requirements.txt
termcolor~=3.1.0
torch
accelerate
ipykernel
ipywidgets
pytest
ipytest
black~=25.1.0
libcst~=1.8.0
requests
control-arena==14.4.1
transformers
protobuf
sentencepiece
mcp[cli]>=1.9.0
PyJWT>=2.9.0
scikit-learn
openai
python-dotenv